feat(app): Allow loading of dns config file from IPFS
The 'dnsConfig' option on start now accepts an IPFS has address. The assumption is that the hash
points at a json file that meets the validation requirements for a DNS config. To help with this
there is a new 'verify' command, that runs a check on the address. The 'verify' command also works
on local files.

Closes #1